feat(wpilib): HAL wrapper
