+++ crates/wpilib-sys/src/bindings.rs
@@ -9,13 +9,10 @@
 #![allow(non_camel_case_types)]
 #![allow(dead_code)]
 #![allow(non_upper_case_globals)]
-#![allow(intra_doc_link_resolution_failure)]
+#![allow(rustdoc::broken_intra_doc_links)]
 #![allow(clippy::useless_transmute)]
 #![allow(clippy::cast_lossless)]
 #![allow(clippy::trivially_copy_pass_by_ref)]
 #![allow(clippy::unreadable_literal)]
-
-use std::ffi;
-use std::fmt;
 
 include!("./hal_bindings.rs");
