+++ crates/wpilib/src/robot/error.rs
@@ -1,0 +1,7 @@
+use thiserror::Error;
+
+#[derive(Error, Debug)]
+pub enum RobotError {
+    #[error("HAL error.")]
+    HALError(#[from] crate::hal::error::HALError),
+}
