--- example.cpp
@@ -1,3 +1,0 @@
-#include <stdint.h>
-
-int main() { return 0; }
