--- examples/Cargo.toml
@@ -1,41 +1,0 @@
-[package]
-name = "examples"
-version = "0.0.0"
-authors = ["Josh Hejna <josh.hejna@gmail.com>"]
-license = "MIT OR Apache-2.0"
-publish = false
-edition = "2021"
-
-[dependencies]
-wpilib = {path = "../crates/wpilib"}
-wpilib-sys = {path = "../crates/wpilib-sys"}
-
-[package.metadata.frc]
-team-number = 114
-# rio-address = "10.1.14.2"
-target-dir = "../target"
-executable-name = "pneumatics"
-
-[[example]]
-name = "custom_usage"
-path = "./custom_usage.rs"
-
-[[example]]
-name = "digital_out"
-path = "./digital_out.rs"
-
-[[example]]
-name = "pdp"
-path = "./pdp.rs"
-
-[[example]]
-name = "pneumatics"
-path = "./pneumatics.rs"
-
-[[example]]
-name = "timed_robot"
-path = "./timed_robot.rs"
-
-[[example]]
-name = "basic"
-path = "./basic"
