+++ examples/README.md
@@ -1,4 +1,8 @@
 # Examples
 
-`cargo build` will build all of the examples. To test them out, use `cargo-frc`.
-Set the `package.metadata.frc.executable-name` key to the test name you would like to try, then run `cargo frc deploy`.
++-----+-----+
+| Name | Description |
++-----+-----+
+| `basic`     | A basic robot.                       |
+| `iterative` | Usage of the `IterativeRobot` trait. |
++-----+-----+
