+++ examples/basic/src/main.rs
@@ -1,40 +1,14 @@
-extern crate wpilib;
+//! Starts the robot, prints "Hello, world!", and stays in a loop forever.
 
-#[derive(Debug)]
-struct Robot {}
+use wpilib::{
+    hal::{UninitializedHAL, HAL},
+    robot::error::RobotError,
+};
 
-impl wpilib::IterativeRobot for Robot {
-    fn new(_ds: &wpilib::ds::DriverStation) -> Robot {
-        Robot {}
-    }
+fn main() -> Result<(), RobotError> {
+    let _hal: HAL = UninitializedHAL::default().initialize()?;
 
-    fn disabled_init(&mut self) {
-        println!("Transitioning to disabled.");
-    }
+    println!("Hello, world!");
 
-    fn autonomous_init(&mut self) {
-        println!("Transitioning to autonomous mode.");
-    }
-
-    fn teleop_init(&mut self) {
-        println!("Transitioning to teleoperated mode.");
-    }
-
-    fn test_init(&mut self) {
-        println!("Transitioning to test mode.");
-    }
-
-    fn disabled_periodic(&mut self) {}
-
-    fn autonomous_periodic(&mut self) {}
-
-    fn teleop_periodic(&mut self) {}
-
-    fn test_periodic(&mut self) {}
-
-    fn robot_periodic(&mut self) {}
-}
-
-fn main() {
-    wpilib::start_timed::<Robot>()
+    loop {}
 }
