+++ examples/iterative/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "basic"
+name = "iterative"
 version = "0.0.0"
 edition = "2021"
 publish = false
